+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 See [Upgrading] for details how to upgrade.
 
+- Drop loading deprecated `ldap.php`, #899
+
 [3.9.3]: https://github.com/eventum/eventum/compare/v3.9.2...master
 
 ## [3.9.2] - 2020-07-26